Finishes stock_picker in ruby from Odin project
as if it had came to me in a dream, I was thinking about how to solve this problem before going to sleep and trying to thought out the whole logic
in my head before trying to code again. I realized that I could just bruteforce everything and just take the correct answers by only recording an answer
if the index from the sell array is higher than the buy array, without having to worry where I was in the array. I thought about it for a whole night and day,
wrote and tested in 30 minutes.